Remove FreeRTOS and related middleware from project
All FreeRTOS source and header files, as well as related include paths and references, have been deleted from the project. Project configuration files were updated to exclude FreeRTOS, and code was refactored to remove dependencies on the RTOS.

+++ Core/Src/led_indicator.c
@@ -85,7 +85,7 @@ uint32_t led_ms_to_timer_ticks(uint32_t ms) {
  * @return 1 if time has expired, 0 if not
  */
 uint8_t led_time_expired(led_t *led) {
-    uint16_t current_time = __HAL_TIM_GET_COUNTER(led->htim_delay);
+    uint32_t current_time = __HAL_TIM_GET_COUNTER(led->htim_delay);
     if (led->start_time <= led->end_time) {
         // Normal case: no rollover
         if (current_time >= led->end_time) {
